requirements.txt
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
frozendict==2.4.6
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
multitasking==0.0.11
newsapi-python==0.2.7
nltk==3.9.1
numpy==2.2.3
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.6
python-dateutil==2.9.0.post0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.38
ta==0.11.0
textblob==0.19.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
yfinance==0.2.54
